pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdmemory"
version = "0.1.0"
description = "Markdown-first, LLM-driven memory framework organized into a hierarchical Knowledge Tree"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Contributors" }
]
dependencies = [
    "python-frontmatter>=1.0.0",
    "litellm>=1.0.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mdmemory"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
